requirements.txt
Django>=5.2
djangorestframework
django-environ
python-dotenv
supabase
django-cors-headers
